requirements.txt
aiogram==3.4.1
playwright==1.40.0
python-dotenv==1.0.0
PyYAML==6.0.1
aiofiles==23.2.1
aiohttp==3.9.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
